move batcher to its own package
